LangChainWithTextFile/requirements.txt
streamlit
langchain
openai
python-dotenv
pinecone-client
streamlit-chat
chromadb
tiktoken